refactor: Move Home-page
