+++ interfaces/Portal/src/app/app-routing.module.ts
@@ -31,7 +31,7 @@ const routes: Routes = [
   {
     path: AppRoutes.home,
     loadChildren: () =>
-      import('./home/home.module').then((m) => m.HomePageModule),
+      import('./pages/home/home.module').then((m) => m.HomePageModule),
     canActivate: [AuthGuard],
   },
   {
